Livrable 2

+++ backend/src/common/orm.js
@@ -1,17 +1,17 @@
 const Sequelize = require('sequelize');
 const logger = require('./logger');
-const config = require('./config')
+const config = require('config')
 const errorManager  = require('./errors')
 
 var sequelize = new Sequelize({
-  username : config.getSqlCredentialsConfig('user'),
-  password : config.getSqlCredentialsConfig('password'),
-  database : config.getSqlConfig('database'),
-  host : config.getSqlConfig('host'),
-  port : config.getSqlConfig('port'),
+  username : config.get('sql')['credentials']['user'],
+  password : config.get('sql')['credentials']['password'],
+  database : config.get('sql')['database'],
+  host : config.get('sql')['host'],
+  port : config.get('sql')['post'],
   options : {
     retry : {
-      max : config.getSqlConfig('maxRetries'),
+      max : config.get('sql')['maxRetries'],
       },
     },
   dialect: 'mysql',
@@ -41,13 +41,28 @@ exports.findAll = (model, res, attributes) => {
  });
 }
 
-exports.find = (model, res, attributes, functionUpdate) => {
+exports.findAllAndCount = (model, res, attributes, attributesCount) => {
+  model.findAndCountAll(attributes
+  ).then(function(result) {
+    if (!result)
+      res.sendStatus(404);
+    model.count(attributesCount).then(function (resultCount) {
+      result['count'] = resultCount;
+      res.json(result);
+    })
+ }).catch(function(err) {
+    logger.error(err.message);
+    res.status(500).send(err.message);
+ });
+}
+
+exports.find = (model, res, errCode, attributes, functionUpdate) => {
   model.find(attributes
   ).then(function(result) {
-   if (!result)
-    res.sendStatus(404)
-  else if (functionUpdate != undefined)
+  if (functionUpdate != undefined)
     functionUpdate(result, res);
+  else if (!result)
+      res.sendStatus(errCode)
   else
     res.json(result);
  }).catch(function(err) {
@@ -56,7 +71,7 @@ exports.find = (model, res, attributes, functionUpdate) => {
  });
 }
 
-exports.build = (model, res, attributes, callbacks) => {
+exports.build = (model, res, attributes, callbacks, user) => {
   model.build(attributes).save()
   .then(function(result) {
     res.status(201);
@@ -64,7 +79,7 @@ exports.build = (model, res, attributes, callbacks) => {
       res.send();
     } else {
       for (var i = 0; i < callbacks.length; ++i) {
-        callbacks[i](result.dataValues, attributes, res);
+        callbacks[i](result.dataValues, attributes, res, user);
       }
     }
   }).catch(function(err) {
@@ -73,7 +88,7 @@ exports.build = (model, res, attributes, callbacks) => {
 }
 
 exports.update = (model, newContent, res, attributes, callbacks) => {
-  this.find(model, res, attributes, function(resultModel, res) {
+  this.find(model, res, 404, attributes, function(resultModel, res) {
     resultModel.update(newContent).then(function(result) {
       if (!result)
        res.sendStatus(404)
@@ -90,15 +105,24 @@ exports.update = (model, newContent, res, attributes, callbacks) => {
   });
 }
 
-exports.delete = (model, res, attributes) => {
-  this.find(model, res, attributes, function(resultModel, res) {
-    resultModel.destroy({cascade : true}).then(function(result) {
-      if (!result)
-       res.sendStatus(404)
-     else
-       res.json(result);
-     }).catch(function(err) {
+exports.delete = (model, res, attributes, callback) => {
+  this.find(model, res, 404, attributes, function(resultModel, res) {
+    if (!resultModel)
+      res.sendStatus(404);
+    else {
+      resultModel.destroy({cascade : true}).then(function(result) {
+        if (!result)
+          res.sendStatus(404)
+        else if (callback != undefined)
+          callback(result, res);
+        res.json(result);
+      }).catch(function(err) {
        errorManager.handle(err, res);
      });
+   }
   });
+}
+
+exports.query = (request) => {
+  sequelize.query(request);
 }
